+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3] - Feb 4, 2023
+
+### Changed
+
+- Remove `chrome-aws-lambda` and `puppeteer`, go back to using only `Cheerio`
+- Codeberg link in the FAQ and the Codeberg logo in the footer/header
+
 ## [2.0.2] - Feb 3, 2023
 
 ### Changed